split combat.js into logic / render / events + fix review bugs
- Split 2775-line monolith into combat-logic.js (state + mechanics),
  combat-render.js (HTML generation + screens), combat-events.js
  (event binding), combat.js (11-line entry point). Circular render↔events
  dependency resolved via _R callback object registered at module load time.

- Fix: handleActivateDomain equal-refinement clash now spends spell slots,
  domain charge, and action before returning — attacker no longer suppresses
  enemy domain for free.
- Fix: splash fumble counter-attack moved outside the hit guard so a nat-1
  on a splash ability correctly grants the defender a free counter-attack.
- Fix: cone aim-mode drawMap reduced from 3 full grid sweeps per mousemove
  to 2 by deriving hit targets from the already-computed cells set instead
  of calling coneCellTargets again.
- Fix: ability fire-button label extracted as exported abilityFireLabel()
  shared by both actionPanelHTML and syncAbilityBtn — label is now a single
  source of truth.
- Cleanup: removed unused SPELL_RANGE import from combat-events.js,
  dead hasAtkTarget / MELEE_RANGE from combat-render.js, and the
  never-called updateAbilityTargetDropdown / targetOptsForAbilityDynamic
  exports.
